fix: don't know why fly don't work

+++ Project4_VirtualMemory/kernel/loader/loader.c
@@ -14,13 +14,13 @@ void load_task_img(int task_id, uintptr_t kva, uintptr_t prev_kva, int page_num)
      * 2. [p1-task4] load task via task name, thus the arg should be 'char *taskname'
      */
 
-    int filesz = tasks[task_id].filesz;
-    int memsz = tasks[task_id].memsz;
-    int offset = tasks[task_id].start_addr % BLOCK_SIZE;
-    int block_id = tasks[task_id].start_addr / BLOCK_SIZE + page_num * BLOCK_NUM;
+    uint32_t filesz = tasks[task_id].filesz;
+    uint32_t memsz = tasks[task_id].memsz;
+    uint32_t offset = tasks[task_id].start_addr % BLOCK_SIZE;
+    uint32_t block_id = tasks[task_id].start_addr / BLOCK_SIZE + page_num * BLOCK_NUM;
 
     // bss sector
-    if ((int)(page_num * NORMAL_PAGE_SIZE - offset) > (int)filesz)
+    if ((int64_t)(page_num * NORMAL_PAGE_SIZE - offset) > (int64_t)filesz)
         return;
 
     bios_sdread(kva2pa(kva), BLOCK_NUM, block_id);
