+++ Project4_VirtualMemory/kernel/sched/sched.c
@@ -51,7 +51,7 @@ pid_t process_id = 1;
 
 pcb_t* list_to_pcb(list_node_t *list)
 {
-    return (pcb_t*)((int)list-2*sizeof(reg_t)-2*sizeof(ptr_t));
+    return (pcb_t*)((uint64_t)list-2*sizeof(reg_t)-2*sizeof(ptr_t));
 }
 
 void do_scheduler(void)
